chore(utils/game): restructure utility files into folders for add unit tests
