+++ src/utils/game/checkWinner/checkWinner.ts
