+++ src/utils/game/getScores/getScores.ts
