+++ src/utils/game/getStatus/getStatus.ts
