+++ src/utils/game/index.ts
@@ -1,8 +1,8 @@
-export * from "./checkMovesLeft";
-export * from "./checkWinner";
-export * from "./findBestMove";
-export * from "./getScores";
-export * from "./getStatus";
-export * from "./makeMove";
-export * from "./minimax";
-export * from "./resetBoard";
+export * from "./checkMovesLeft/checkMovesLeft";
+export * from "./checkWinner/checkWinner";
+export * from "./findBestMove/findBestMove";
+export * from "./getScores/getScores";
+export * from "./getStatus/getStatus";
+export * from "./makeMove/makeMove";
+export * from "./minimax/minimax";
+export * from "./resetBoard/resetBoard";
