+++ src/utils/game/minimax/minimax.ts
