+++ src/utils/game/resetBoard/resetBoard.ts
